+++ bingads/v13/bulk/entities/audiences/__init__.py
@@ -1,10 +1,12 @@
 __author__ = 'Bing Ads SDK Team'
 __email__ = 'bing_ads_sdk@microsoft.com'
 
+from .bulk_audience import *
 from .bulk_remarketing_list import *
 from .bulk_custom_audience import *
 from .bulk_ad_group_custom_audience_association import *
 from .bulk_in_market_audience import *
+from .bulk_combined_list import *
 from .bulk_ad_group_in_market_audience_association import *
 from .bulk_ad_group_product_audience_association import *
 from .bulk_ad_group_negative_product_audience_association import *
@@ -12,7 +14,9 @@ from .bulk_ad_group_negative_in_market_audience_association import *
 from .bulk_ad_group_negative_custom_audience_association import *
 from .bulk_ad_group_negative_remarketing_list_association import *
 from .bulk_ad_group_negative_similar_remarketing_list_association import *
+from .bulk_ad_group_negative_combined_list_association import *
 from .bulk_ad_group_remarketing_list_association import *
+from .bulk_ad_group_combined_list_association import *
 from .bulk_product_audience import *
 from .bulk_similar_remarketing_list import *
 from .bulk_ad_group_similar_remarketing_list_association import *
@@ -24,9 +28,17 @@ from .bulk_campaign_in_market_audience_association import *
 from .bulk_campaign_product_audience_association import *
 from .bulk_campaign_remarketing_list_association import *
 from .bulk_campaign_similar_remarketing_list_association import *
+from .bulk_campaign_combined_list_association import *
 from .bulk_campaign_negative_audience_association import *
 from .bulk_campaign_negative_remarketing_list_association import *
 from .bulk_campaign_negative_custom_audience_association import *
 from .bulk_campaign_negative_in_market_audience_association import *
 from .bulk_campaign_negative_product_audience_association import *
 from .bulk_campaign_negative_similar_remarketing_list_association import *
+from .bulk_campaign_negative_combined_list_association import *
+from .bulk_customer_list import *
+from .bulk_ad_group_customer_list_association import *
+from .bulk_ad_group_negative_customer_list_association import *
+from .bulk_campaign_customer_list_association import *
+from .bulk_campaign_negative_customer_list_association import *
+from .bulk_customer_list_item import *