+++ bingads/v13/bulk/entities/audiences/bulk_ad_group_customer_list_association.py
@@ -1,0 +1,14 @@
+from bingads.v13.bulk.entities.audiences.bulk_ad_group_audience_association import BulkAdGroupAudienceAssociation
+
+class BulkAdGroupCustomerListAssociation(BulkAdGroupAudienceAssociation):
+    """ Represents an Ad Group Customer List Association that can be read or written in a bulk file.
+
+    For more information, see Ad Group Customer List Association at https://go.microsoft.com/fwlink/?linkid=846127.
+
+    *See also:*
+
+    * :class:`.BulkServiceManager`
+    * :class:`.BulkOperation`
+    * :class:`.BulkFileReader`
+    * :class:`.BulkFileWriter`
+    """