+++ bingads/v13/bulk/entities/audiences/bulk_ad_group_negative_combined_list_association.py
@@ -1,0 +1,15 @@
+from bingads.v13.bulk.entities.audiences.bulk_ad_group_negative_audience_association import *
+
+
+class BulkAdGroupNegativeCombinedListAssociation(BulkAdGroupNegativeAudienceAssociation):
+    """ Represents an Ad Group Negative Combined list Association that can be read or written in a bulk file.
+
+    For more information, see Ad Group Negative Combined List Association at https://go.microsoft.com/fwlink/?linkid=846127.
+
+    *See also:*
+
+    * :class:`.BulkServiceManager`
+    * :class:`.BulkOperation`
+    * :class:`.BulkFileReader`
+    * :class:`.BulkFileWriter`
+    """