+++ bingads/v13/bulk/entities/audiences/bulk_campaign_customer_list_association.py
@@ -1,0 +1,14 @@
+from bingads.v13.bulk.entities.audiences.bulk_campaign_audience_association import BulkCampaignAudienceAssociation
+
+class BulkCampaignCombinedListAssociation(BulkCampaignAudienceAssociation):
+    """ Represents an Campaign Customer List Association that can be read or written in a bulk file.
+
+    For more information, see Campaign Customer List Association at https://go.microsoft.com/fwlink/?linkid=846127.
+
+    *See also:*
+
+    * :class:`.BulkServiceManager`
+    * :class:`.BulkOperation`
+    * :class:`.BulkFileReader`
+    * :class:`.BulkFileWriter`
+    """