+++ bingads/v13/bulk/entities/audiences/bulk_campaign_negative_combined_list_association.py
@@ -1,0 +1,15 @@
+from bingads.v13.bulk.entities.audiences.bulk_campaign_negative_audience_association import *
+
+
+class BulkCampaignNegativeCombinedListAssociation(BulkCampaignNegativeAudienceAssociation):
+    """ Represents an Campaign Negative Combined List Association that can be read or written in a bulk file.
+
+    For more information, see Campaign Negative Combined List Association at https://go.microsoft.com/fwlink/?linkid=846127.
+
+    *See also:*
+
+    * :class:`.BulkServiceManager`
+    * :class:`.BulkOperation`
+    * :class:`.BulkFileReader`
+    * :class:`.BulkFileWriter`
+    """