+++ bingads/v13/bulk/entities/audiences/bulk_remarketing_list.py
@@ -4,8 +4,9 @@ from bingads.v13.internal.bulk.entities.single_record_bulk_entity import _Single
 from bingads.v13.internal.bulk.mappings import _SimpleBulkMapping
 from bingads.v13.internal.bulk.string_table import _StringTable
 from bingads.v13.internal.extensions import *
+from .bulk_audience import BulkAudience
 
-class BulkRemarketingList(_SingleRecordBulkEntity):
+class BulkRemarketingList(BulkAudience):
     """ Represents an Remarketing List that can be read or written in a bulk file.
 
     This class exposes the :attr:`remarketing_list` property that can be read and written as fields of the
@@ -24,47 +25,10 @@ class BulkRemarketingList(_SingleRecordBulkEntity):
     def __init__(self,
                  remarketing_list=None,
                  status=None,):
-        super(BulkRemarketingList, self).__init__()
-
-        self._remarketing_list = remarketing_list
-        self._status = status
+        super(BulkRemarketingList, self).__init__(audience = remarketing_list, status = status)
 
     _MAPPINGS = [
-        _SimpleBulkMapping(
-            header=_StringTable.Status,
-            field_to_csv=lambda c: c.status,
-            csv_to_field=lambda c, v: setattr(c, 'status', v)
-        ),
-        _SimpleBulkMapping(
-            _StringTable.Id,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.Id),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'Id', int(v) if v else None)
-        ),
-        _SimpleBulkMapping(
-            _StringTable.ParentId,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.ParentId),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'ParentId', int(v) if v else None)
-        ),
-        _SimpleBulkMapping(
-            _StringTable.Audience,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.Name),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'Name', v)
-        ),
-        _SimpleBulkMapping(
-            _StringTable.Description,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.Description),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'Description', v)
-        ),
-        _SimpleBulkMapping(
-            _StringTable.MembershipDuration,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.MembershipDuration),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'MembershipDuration', int(v) if v else None)
-        ),
-        _SimpleBulkMapping(
-            header=_StringTable.Scope,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.Scope),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'Scope', v if v else None)
-        ),
+       
         _SimpleBulkMapping(
             _StringTable.TagId,
             field_to_csv=lambda c: bulk_str(c.remarketing_list.TagId),
@@ -75,54 +39,26 @@ class BulkRemarketingList(_SingleRecordBulkEntity):
             field_to_csv=lambda c: field_to_csv_RemarketingRule(c.remarketing_list),
             csv_to_field=lambda c, v: csv_to_field_RemarketingRule(c.remarketing_list, v)
         ),
-        _SimpleBulkMapping(
-            _StringTable.AudienceSearchSize,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.SearchSize),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'SearchSize', int(v) if v else None)
-        ),       
-        _SimpleBulkMapping(
-            header=_StringTable.AudienceNetworkSize,
-            field_to_csv=lambda c: bulk_str(c.remarketing_list.AudienceNetworkSize),
-            csv_to_field=lambda c, v: setattr(c.remarketing_list, 'AudienceNetworkSize', v if v else None)
-        ),
-        _SimpleBulkMapping(
-            _StringTable.SupportedCampaignTypes,
-            field_to_csv=lambda c: field_to_csv_SupportedCampaignTypes(c.remarketing_list.SupportedCampaignTypes),
-            csv_to_field=lambda c, v: csv_to_field_SupportedCampaignTypes(c.remarketing_list.SupportedCampaignTypes, v)
-        ),
     ]
 
     @property
     def remarketing_list(self):
         """ Defines a Remarketing List """
 
-        return self._remarketing_list
+        return self._audience
 
     @remarketing_list.setter
     def remarketing_list(self, remarketing_list):
-        self._remarketing_list = remarketing_list
-
-    @property
-    def status(self):
-        """ The status of the Remarketing List
-
-        :rtype: str
-        """
-
-        return self._status
-
-    @status.setter
-    def status(self, status):
-        self._status = status
-
+        self._audience = remarketing_list
+   
 
     def process_mappings_to_row_values(self, row_values, exclude_readonly_data):
         self._validate_property_not_null(self.remarketing_list, 'remarketing_list')
+        super(BulkRemarketingList, self).process_mappings_to_row_values(row_values, exclude_readonly_data)
         self.convert_to_values(row_values, BulkRemarketingList._MAPPINGS)
 
     def process_mappings_from_row_values(self, row_values):
-        self._remarketing_list = _CAMPAIGN_OBJECT_FACTORY_V13.create('RemarketingList')
+        self.remarketing_list = _CAMPAIGN_OBJECT_FACTORY_V13.create('RemarketingList')
+        super(BulkRemarketingList, self).process_mappings_from_row_values(row_values)
         row_values.convert_to_entity(self, BulkRemarketingList._MAPPINGS)
 
-    def read_additional_data(self, stream_reader):
-        super(BulkRemarketingList, self).read_additional_data(stream_reader)